pyproject.toml
[project]
name = "asr-whisperx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.120.3",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
    "whisperx",
    "matplotlib>=3.10.7",
    "pyannote-audio[separation]==3.3.2",
    "torchaudio==2.8.0",
    "torchvision>=0.23.0",
]

[tool.uv.sources]
whisperx = { git = "https://github.com/m-bain/whisperX.git" }
